Remove Binderpos search API dead code and fallback comments
