+++ api/gateway/cardscitadel/search.go
@@ -10,8 +10,6 @@ const StoreName = "Cards Citadel"
 const StoreBaseURL = "https://cardscitadel.com"
 const StoreSearchURL = "/search?q=*%s*"
 
-const binderposStoreURL = "card-citadel.myshopify.com"
-
 type Store struct {
 	Name         string
 	BaseUrl      string
@@ -30,22 +28,4 @@ func NewLGS() gateway.LGS {
 
 func (s Store) Search(ctx context.Context, searchStr string) ([]gateway.Card, error) {
 	return s.BinderposGwy.Scrap(ctx, 1, s.Name, s.BaseUrl, s.SearchUrl, searchStr)
-
-	// reqPayload, err := json.Marshal(binderpos.Payload{
-	// 	StoreURL:    binderposStoreURL,
-	// 	Game:        binderpos.ProductTypeMTG.ToString(),
-	// 	Title:       searchStr,
-	// 	InstockOnly: true,
-	// })
-	// if err != nil {
-	// 	return []gateway.Card{}, err
-	// }
-
-	// cards, httpStatusCode, err := s.BinderposGwy.Search(ctx, s.Name, s.BaseUrl, reqPayload)
-	// if err != nil || httpStatusCode != http.StatusOK {
-	// 	log.Printf("falling back to scrap for [%s]", s.Name)
-	// 	return s.BinderposGwy.Scrap(ctx, 1, s.Name, s.BaseUrl, s.SearchUrl, searchStr)
-	// }
-
-	// return cards, nil
 }
