+++ api/gateway/flagship/search.go
@@ -11,8 +11,6 @@ const StoreName = "Flagship Games"
 const StoreBaseURL = "https://www.flagshipgames.sg"
 const StoreSearchURL = "/search?type=product&q=%s"
 
-const binderposStoreURL = "flagship-games.myshopify.com"
-
 type Store struct {
 	Name         string
 	BaseUrl      string
@@ -31,22 +29,4 @@ func NewLGS() gateway.LGS {
 
 func (s Store) Search(ctx context.Context, searchStr string) ([]gateway.Card, error) {
 	return s.BinderposGwy.Scrap(ctx, 2, s.Name, s.BaseUrl, s.SearchUrl, searchStr)
-
-	// reqPayload, err := json.Marshal(binderpos.Payload{
-	// 	StoreURL:    binderposStoreURL,
-	// 	Game:        binderpos.ProductTypeMTG.ToString(),
-	// 	Title:       searchStr,
-	// 	InstockOnly: true,
-	// })
-	// if err != nil {
-	// 	return []gateway.Card{}, err
-	// }
-
-	// cards, httpStatusCode, err := s.BinderposGwy.Search(ctx, s.Name, s.BaseUrl, reqPayload)
-	// if err != nil || httpStatusCode != http.StatusOK {
-	// 	log.Printf("falling back to scrap for [%s]", s.Name)
-	// 	return s.BinderposGwy.Scrap(ctx, 2, s.Name, s.BaseUrl, s.SearchUrl, searchStr)
-	// }
-
-	// return cards, nil
 }
