+++ api/gateway/onemtg/search.go
@@ -10,8 +10,6 @@ const StoreName = "OneMtg"
 const StoreBaseURL = "https://onemtg.com.sg"
 const StoreSearchURL = "/search?q=%s"
 
-const binderposStoreURL = "one-mtg.myshopify.com"
-
 type Store struct {
 	Name         string
 	BaseUrl      string
@@ -36,28 +34,4 @@ func (s Store) Search(ctx context.Context, searchStr string) ([]gateway.Card, er
 		s.SearchUrl,
 		searchStr,
 	)
-
-	// reqPayload, err := json.Marshal(binderpos.Payload{
-	// 	StoreURL:    binderposStoreURL,
-	// 	Game:        binderpos.ProductTypeMTG.ToString(),
-	// 	Title:       searchStr,
-	// 	InstockOnly: true,
-	// })
-	// if err != nil {
-	// 	return []gateway.Card{}, err
-	// }
-
-	// cards, httpStatusCode, err := s.BinderposGwy.Search(ctx, s.Name, s.BaseUrl, reqPayload)
-	// if err != nil || httpStatusCode != http.StatusOK {
-	// 	log.Printf("falling back to scrap for [%s]", s.Name)
-	// 	return s.BinderposGwy.Scrap(ctx,
-	// 		2,
-	// 		s.Name,
-	// 		s.BaseUrl,
-	// 		s.SearchUrl,
-	// 		searchStr,
-	// 	)
-	// }
-
-	// return cards, nil
 }
